requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
docutils==0.18.1
    # via kivy
idna==3.3
    # via requests
kivy==2.1.0
    # via
    #   -r requirements.in
    #   kivymd
kivy-garden==0.1.5
    # via kivy
kivymd==0.104.2
    # via -r requirements.in
pillow==9.2.0
    # via kivymd
pygments==2.12.0
    # via kivy
pyserial==3.5
    # via -r requirements.in
pyyaml==6.0
    # via -r requirements.in
requests==2.28.1
    # via kivy-garden
urllib3==1.26.9
    # via requests
